Add ability to read files and output random lines
Introduce a internal name space for internal procedures.

Add a procedure for parsing files into tcl lists and two procedures for
outputting ether a random index or a random element from a list.

These new procedures combined allows the bot to print a random line from
a file in a scalable way.

Also, extent the offline test with this functionality.

Signed-off-by: Richard A (rical) <ric41@yahoo.se>