Make db and txt parsing working directory agnostic
This includes removing the sanity check in parse_file. It would have
failed as the paths now are absolute. Rather then trying to be clever
and taking hight for this I simply remove it. The file names should never
be passed by the user and we should have control over there names
(famous last words).

Signed-off-by: Richard A (rical) <rical@epoch70.com>