Add AI functionality (give the bot a brain)
This is a very simple artificial intelligence attempt that learns
opinions about a topic based on a keyword. This keyword is currently the
Swedish word for "is".

Learning from marker "is":
user1: Sweden is to cold!
Adds the opinion "is to cold!" to the topic "Sweden".

Talking about topic "Sweden":
user2: I'm thinking about going to Sweden
bot: user2: Sweden is to cold!

I assume this theory is full of loopholes and sentences that doesn't
make sense. To address this I added the !stupid functionality.

Issuing "!stupid simply not true" in the chat will make the bot
remove the last statement (talk) and record info about the removal.

Signed-off-by: Richard A (rical) <rical@epoch70.com>