friday: remove Friday concept
Quote printing, which was the only thing using friday mode has been
disabled for a long period of time. Today friday-mode is primarily the
AI talking which isn't using the friday framework.

In the future we might want to reimplement some of the functions from
friday mode, for example the backoff to make the AI talking less
aggressive. However, I feel having a "friday framework" isn't really
justifiable at this point.