ai: don't learn about "badwords" only hand out streck
This is not a feature. Ideally the user should get a streck and then
we should learn, however the current design prohibits that without
"major" changes.

So in this patch we change the behavior to give a streck but never
learn about it. Previously one could avoid getting a streck by
formulating the sentence in a way that triggered the AI learning.